+++ NEWS.md
@@ -5,6 +5,8 @@
 - `get_users()` now supports filtering users with the `account_status` and
   `user_role` parameters. For example, this allows you to find all licensed
   users on a Connect server. (#311)
+- The new `get_group_content()` function lets you view the content that groups
+  have permission to access. (#334)
 
 # connectapi 0.4.0